plugin namespace. rename plugin. move test resources to test CAT.
